+++ cmd/symbols/main.go
@@ -34,7 +34,6 @@ var (
 	cacheDir       = env.Get("CACHE_DIR", "/tmp/symbols-cache", "directory to store cached symbols")
 	cacheSizeMB    = env.Get("SYMBOLS_CACHE_SIZE_MB", "100000", "maximum size of the disk cache in megabytes")
 	ctagsProcesses = env.Get("CTAGS_PROCESSES", strconv.Itoa(runtime.NumCPU()), "number of ctags child processes to run")
-	ctagsCommand   = env.Get("CTAGS_COMMAND", "universal-ctags", "ctags command (should point to universal-ctags executable compiled with JSON and seccomp support)")
 )
 
 const port = "3184"
@@ -45,6 +44,8 @@ func main() {
 	log.SetFlags(0)
 	tracer.Init()
 
+	symbols.MustRegisterSqlite3WithPcre()
+
 	go debugserver.Start()
 
 	service := symbols.Service{
@@ -52,9 +53,9 @@ func main() {
 			return git.Archive(ctx, repo, git.ArchiveOptions{Treeish: string(commit), Format: "tar"})
 		},
 		NewParser: func() (ctags.Parser, error) {
-			parser, err := ctags.NewParser(ctagsCommand)
+			parser, err := ctags.NewParser(ctags.GetCommand())
 			if err != nil {
-				return nil, errors.Wrap(err, fmt.Sprintf("command: %s", ctagsCommand))
+				return nil, errors.Wrap(err, fmt.Sprintf("command: %s", ctags.GetCommand()))
 			}
 			return parser, nil
 		},