+++ dev/go-install.sh
@@ -41,20 +41,6 @@ case $# in
 	done
 	;;
 esac
-if [ ! -z "$ENTERPRISE_COMMANDS" ]; then
-    for entCmd in $ENTERPRISE_COMMANDS; do
-        exists=false
-        for cmd in $commands; do
-            if [ "$cmd" = "$entCmd" ]; then
-                exists=true
-                break
-            fi
-        done
-        if [ "$exists" = false ]; then
-            commands="$commands $entCmd "
-        fi
-    done
-fi
 
 $ok || exit 1
 
@@ -70,7 +56,7 @@ github.com/google/zoekt/cmd/zoekt-webserver \
 "
 
 if [ ! -n "${OFFLINE-}" ]; then
-    INSTALL_GO_PKGS="$INSTALL_GO_PKGS github.com/derekparker/delve/cmd/dlv"
+    INSTALL_GO_PKGS="$INSTALL_GO_PKGS github.com/go-delve/delve/cmd/dlv"
 fi
 
 if ! go install $INSTALL_GO_PKGS; then