+++ dev/launch.sh
@@ -51,8 +51,6 @@ export REPO_UPDATER_URL=http://127.0.0.1:3182
 export REDIS_ENDPOINT=127.0.0.1:6379
 export QUERY_RUNNER_URL=http://localhost:3183
 export SYMBOLS_URL=http://localhost:3184
-export CTAGS_COMMAND=${CTAGS_COMMAND-cmd/symbols/universal-ctags-dev}
-export CTAGS_PROCESSES=1
 export SRC_SYNTECT_SERVER=http://localhost:9238
 export SRC_FRONTEND_INTERNAL=localhost:3090
 export SRC_PROF_HTTP=
@@ -83,11 +81,6 @@ yarn_pid=''
     yarn_pid="$!"
 }
 
-# Build the symbols image
-env IMAGE=dev-symbols ./cmd/symbols/build.sh >/dev/null
-symbols_pid="$!"
-rm -rf /tmp/symbols-cache
-
 if ! ./dev/go-install.sh; then
 	# let Yarn finish, otherwise we get Yarn diagnostics AFTER the
 	# actual reason we're failing.
@@ -100,9 +93,6 @@ fi
 if [[ -n "$yarn_pid" ]]; then
     wait "$yarn_pid"
 fi
-
-# Wait for the symbols image to finish building
-wait "$symbols_pid"
 
 # Increase ulimit (not needed on Windows/WSL)
 type ulimit > /dev/null && ulimit -n 10000 || true