+++ enterprise/dev/ci/gen-pipeline.go
@@ -92,7 +92,8 @@ func main() {
 			bk.Cmd("./dev/ci/ci-db-backcompat.sh"))
 
 		pipeline.AddStep(":go:",
-			bk.Cmd("go generate ./..."),
+			bk.Cmd("yarn"),
+			bk.Cmd("./cmd/symbols/build.sh buildLibsqlite3Pcre"), // for symbols tests
 			bk.Cmd("go test -coverprofile=coverage.txt -covermode=atomic -race ./..."),
 			bk.ArtifactPaths("coverage.txt"))
 
@@ -133,11 +134,21 @@ func main() {
 		}
 
 		image := "sourcegraph/" + app
-		buildScript := cmdDir + "/build.sh"
+
+		getBuildScript := func() string {
+			buildScriptByApp := map[string]string{
+				"symbols": "env BUILD_TYPE=dist ./cmd/symbols/build.sh buildSymbolsDockerImage",
+			}
+			if buildScript, ok := buildScriptByApp[app]; ok {
+				return buildScript
+			}
+			return cmdDir + "/build.sh"
+		}
+
 		cmds = append(cmds,
 			bk.Env("IMAGE", image+":"+version),
 			bk.Env("VERSION", version),
-			bk.Cmd(buildScript),
+			bk.Cmd(getBuildScript()),
 		)
 
 		if app != "server" || taggedRelease {