+++ CHANGELOG.md
@@ -11,11 +11,23 @@ All notable changes to Sourcegraph are documented in this file.
 
 ### Changed
 
+- Symbols search is much faster now. After the initial indexing, you can expect code intelligence to be nearly instant no matter the size of your repository.
+
 ### Fixed
 
 ### Removed
 
-## 3.1
+## 3.1.1
+
+### Added
+
+- Query builder toggle (open/closed) state is now retained.
+
+### Fixed
+
+- Fixed an issue where single-term values entered into the "Exact match" field in the query builder were not getting wrapped in quotes.
+
+## 3.1.0
 
 ### Added
 
@@ -23,12 +35,13 @@ All notable changes to Sourcegraph are documented in this file.
 - Added syntax highlighting for Kotlin and Dart.
 - Added a management console environment variable to disable HTTPS, see [the docs](doc/admin/management_console.md#can-i-disable-https-on-the-management-console) for more information.
 - Added `auth.disableUsernameChanges` to critical configuration to prevent users from changing their usernames.
+- Site admins can query a user by email address or username from the GraphQL API.
 - Added a search query builder to the main search page. Click "Use search query builder" to open the query builder, which is a form with separate inputs for commonly used search keywords.
 
 ### Changed
 
-- File match search results now show full repo name if there are results from mirrors on different code hosts (e.g. github.com/sourcegraph/sourcegraph and gitlab.com/sourcegraph/sourcegraph)
-- Search queries now use "smart case" by default. Searches are case insensitive unless you use uppercase letters. To explicitely set the case, you can still use the `case` field (e.g. `case:yes`, `case:no`). To explicitely set smart case, use `case:auto`.
+- File match search results now show full repository name if there are results from mirrors on different code hosts (e.g. github.com/sourcegraph/sourcegraph and gitlab.com/sourcegraph/sourcegraph)
+- Search queries now use "smart case" by default. Searches are case insensitive unless you use uppercase letters. To explicitly set the case, you can still use the `case` field (e.g. `case:yes`, `case:no`). To explicitly set smart case, use `case:auto`.
 
 ### Fixed